Remove bk spm basic and sales funnel Fix Google Tag Manager Add info for next-seo config git commit -m
