+++ next-seo.config.js
@@ -8,9 +8,10 @@ export default {
     site_name: "Sang Le Tech",
     url: "https://sangletech.com",
     title: "Sang Le Tech",
-    description: "SEO made easy for Next.js projects",
+    description:
+      "VƯƠN TẦM BẢN THÂN - Giúp Bạn Bước Đi Tự Tin Trên Hành Trình Làm Giàu 4.0",
     image:
-      "https://prismic-io.s3.amazonaws.com/gary-blog%2F3297f290-a885-4cc6-9b19-3235e3026646_default.jpg",
+      "https://super-static-assets.s3.amazonaws.com/8090ac7d-2195-4ec8-96be-1457b54a8900/images/d508c9e1-75af-4c14-bcd4-eb63b26317f5.png?w=1500&f=webp",
     imageWidth: 800,
     imageHeight: 600,
   },
